changing front #2

+++ script.js
@@ -12,18 +12,18 @@ window.onload = async () => {
     });
     const result = await resp.json();
     itemList = result;
+
+    render();
   }
-  catch {
+  catch (error) {
     alert('cant get all items');
   }
-
-  render();
 }
 
 const render = () => {
   const list = document.getElementById('list-with-costs');
   if (list === null) {
-    alert('no list');
+    return;
   } 
 
   while (list.firstChild) {
@@ -31,41 +31,56 @@ const render = () => {
   }
 
   const total = document.getElementById('total-sum');
-  total.innerText = 0;
+  let totalSum = itemList.reduce((sum, currentSum) => sum += currentSum.cost, 0);
+  total.innerText = `${totalSum} р.`; 
 
   itemList.forEach((item, index) => {
     const { _id, where, when, cost } = item;
-    const itemSum = /[0-9]+/;
+
+    const now = moment(when).format('DD.MM.YYYY');
 
     const container = document.createElement('div');
-    const writeWhere = document.createElement('p');
-    const writeWhen = document.createElement('p');
+    const containerInfo = document.createElement('div');
+    const position = document.createElement('p');
+    const whereText = document.createElement('p');
+    const whenText = document.createElement('p');
     const sum = document.createElement('p');
 
     list.appendChild(container);
-    
-    container.appendChild(writeWhere);
-
-    container.appendChild(writeWhen);
-    container.appendChild(sum);
+    container.appendChild(position);
+    container.appendChild(whereText);
+    container.appendChild(containerInfo);
+    containerInfo.appendChild(whenText);
+    containerInfo.appendChild(sum);
 
     container.id = `container-${_id}`;
     container.className = 'list__container';
 
-    [writeWhere.innerText, writeWhen.innerText, sum.innerText] = [`${index + 1}) ${where}`, when, `${cost} р.`];
-    [writeWhere.className, writeWhen.className, sum.className] = ['write-where', 'write-when', 'sum'];
+    whereText.id = `where-${_id}`;
+    whenText.id = `when-${_id}`;
+    sum.id = `sum-${_id}`;
 
-    const matchedTotal = total.innerText.match(itemSum);
-    let totalSum = Number(matchedTotal);
-    totalSum += Number(cost);
-    total.innerText = `${totalSum} р.`; 
+    containerInfo.id = `container-info-${_id}`;
+    containerInfo.className = `list__container-info`;
 
-    addButtons(_id, writeWhere.innerText, writeWhen.innerText, sum.innerText);
+    position.innerText = `${index + 1})`;
+    whereText.innerText = where;
+    whenText.innerText = now;
+    sum.innerText = `${cost} p.`;
 
+    whereText.className = 'write-where';
+    whenText.className = 'write-when';
+    sum.className = 'sum';
+
+    addButtons(_id, whereText.innerText, whenText.innerText, sum.innerText);
   });
 }
 
 const addItem = async () => {
+  const myCostsList = document.getElementById(`list-with-costs`);
+  if (myCostsList === null) {
+    return;
+  }
   const inputWhere = document.getElementById('where');
   const inputHowMuch = document.getElementById('howMuch');
 
@@ -75,45 +90,52 @@ const addItem = async () => {
   }
 
   try {
-    const date = new Date();
-    const today = date.toLocaleDateString('en-GB', {
-      day: '2-digit',
-      month: '2-digit',
-      year: 'numeric'
-    });
-
-    const [day, month, year] = today.split('/');
-    const now = day + "." + month + "." + year;
-    
     const resp = await fetch(`${host}/newItem`, {
       method: 'POST',
       headers: hdrs,
       body: JSON.stringify({
         where: inputWhere.value,
-        when: now,
-        cost: inputHowMuch.value
+        cost: Number(inputHowMuch.value)
       })
     });
     const result = await resp.json();
-    itemList = result;
-  }
-  catch(error) {
+    itemList.push(result);
+
+    render();
+  } catch (error) {
     alert('cant add new item');
   }
   
   inputWhere.value = '';
   inputHowMuch.value = '';
-
-  render();
 }
 
-const editItem = async (id, lastWhere, lastWhen, lastSum) => {
-  const itemSum = /[0-9]+/;
-  const index = /[0-9]+[)]/;
-
-  const parent = document.getElementById(`container-${id}`);
-  if (parent === null) {
+const editItem = async (id) => {
+  const myCostsList = document.getElementById(`list-with-costs`);
+  if (myCostsList === null) {
     return;
+  }
+  const itemSum = /[0-9]+/;
+  
+  const container = document.getElementById(`container-${id}`);
+  const containerInfo = document.getElementById(`container-info-${id}`);
+  const buttons = document.getElementById(`buttons-${id}`);
+  const where = document.getElementById(`where-${id}`);
+  const when = document.getElementById(`when-${id}`);
+  const sum = document.getElementById(`sum-${id}`);
+  const editButton = document.getElementById(`edit-button-${id}`);
+  const deleteButton = document.getElementById(`delete-button-${id}`);
+  
+  if (container === null
+    || containerInfo === null
+    || buttons === null
+    || where === null
+    || when === null
+    || sum === null
+    || editButton === null
+    || deleteButton === null
+    ) {
+      return;
   }
 
   const replaceWhere = document.createElement('input');
@@ -124,27 +146,22 @@ const editItem = async (id, lastWhere, lastWhen, lastSum) => {
   const doneImg = document.createElement('img');
   const cancelImg = document.createElement('img');
 
-  const buttons = parent.getElementsByTagName('button');
-  const texts = parent.getElementsByTagName('p');
-
-  const cutOffIndex = texts[0].innerText.match(index);
-
   replaceWhere.className = 'container__replace-where';
-  replaceWhere.value = texts[0].innerText.substring(cutOffIndex[0].length, texts[0].innerText.length).trim();
+  replaceWhere.id = `replace-where-${id}`;
+  replaceWhere.value = where.innerText;
 
+  replaceWhen.type = 'date';
   replaceWhen.className = 'container__replace-when';
-  replaceWhen.value = texts[1].innerText;
+  replaceWhen.id = `replace-when-${id}`;
 
   replaceCost.type = 'number';
   replaceCost.className = 'container__replace-cost';
-  const matchedSum = texts[2].innerText.match(itemSum);
-  replaceCost.value = texts[2].innerText.match(matchedSum[0]);
+  replaceCost.id = `replace-cost-${id}`;
+  const matchedSum = sum.innerText.match(itemSum);
+  replaceCost.value = sum.innerText.match(matchedSum[0]);
 
-  parent.removeChild(buttons[1]);
-  parent.removeChild(buttons[0]);
-  parent.removeChild(texts[2]);
-  parent.removeChild(texts[1]);
-  parent.removeChild(texts[0]);
+  doneButton.id = `done-button-${id}`;
+  cancelButton.id = `cancel-button-${id}`;
 
   doneImg.src = 'img/done.svg';
   cancelImg.src = 'img/cancel.svg';
@@ -152,68 +169,68 @@ const editItem = async (id, lastWhere, lastWhen, lastSum) => {
   doneButton.appendChild(doneImg);
   cancelButton.appendChild(cancelImg);
 
-  parent.appendChild(replaceWhere);
-  parent.appendChild(replaceWhen);
-  parent.appendChild(replaceCost);
-  parent.appendChild(doneButton);
-  parent.appendChild(cancelButton);
-  
+  buttons.appendChild(doneButton);
+  buttons.appendChild(cancelButton);
+
+  container.replaceChild(replaceWhere, where);
+  containerInfo.replaceChild(replaceWhen, when);
+  containerInfo.replaceChild(replaceCost, sum);
+  buttons.replaceChild(doneButton, editButton);
+  buttons.replaceChild(cancelButton, deleteButton);
+
   doneButton.onclick = () => {
     doneItemEditing(id);
   }
 
   cancelButton.onclick = () => {
-    cancelItemEditing(id, lastWhere, lastWhen, lastSum);
+    cancelItemEditing(id, where.innerText, when.innerText, sum.innerText);
   }
 }
 
 const deleteItem = async (id) => {
   try {
+    const myCostsList = document.getElementById(`list-with-costs`);
+    if (myCostsList === null) {
+      return;
+    }
+
     const resp = await fetch(`${host}/deleteItem`, {
       method: 'DELETE',
       headers: hdrs,
       body: JSON.stringify({_id: id})
     });
     const result = await resp.json();
-
+    
     if (result.deletedCount !== 0) {
       itemList = itemList.filter(item => item._id !== id);
     }
-  }
-  catch(error) {
+
+    render();
+  } catch (error) {
     alert('unable to delete task');
   }
-
-  render();
 }
 
 const doneItemEditing = async (id) => {
-  const parent = document.getElementById(`container-${id}`);
-  if (parent === null) {
+  const myCostsList = document.getElementById(`list-with-costs`);
+  if (myCostsList === null) {
     return;
-  }
+  } 
+
+  const changedWhere = document.getElementById(`replace-where-${id}`);
+  const changedWhen = document.getElementById(`replace-when-${id}`);
+  const changedCost = document.getElementById(`replace-cost-${id}`);
 
   const itemSum = /[0-9]+/;
-  const itemNewDate = /^\d{2}\.\d{2}\.\d{4}$/;
 
-  const changedValues = parent.getElementsByTagName('input');
-
-  const changedWhere = changedValues[0].value;
-  const changedWhen = changedValues[1].value;
-  const changedCost = changedValues[2].value;
-
-  if (changedWhere.trim() === '' || changedWhen.trim() === '' || changedCost.trim() === '') {
+  if (changedWhere.value.trim() === ''
+    || changedWhen.value.trim() === '' 
+    || changedCost.value.trim() === ''
+  ) {
     return;
   }
 
-  const matchedSum = changedCost.match(itemSum);
-
-  const [writeWhere, writeWhen, sum] = [ `${changedWhere}`, changedWhen, `${changedCost.match(matchedSum[0])}`];
-
-  if (!changedWhen.match(itemNewDate)) {
-    changedValues[1].focus();
-    return;
-  }
+  const matchedSum = changedCost.value.match(itemSum);
 
   try {
     const resp = await fetch(`${host}/editItem`, {
@@ -221,82 +238,95 @@ const doneItemEditing = async (id) => {
       headers: hdrs,
       body: JSON.stringify({
         _id: id,
-        where: writeWhere,
-        when: writeWhen,
-        cost: sum 
+        where: changedWhere.value,
+        when: changedWhen.value,
+        cost: `${changedCost.value.match(matchedSum[0])}`
       })
     });
     const result = await resp.json();
-    itemList = result;
-  }
-  catch(error) {
+    result.when = moment(result.when).format('MM.DD.YYYY');
+
+    for (let i = 0; i < itemList.length; i++) {
+      if (itemList[i]._id === id) {
+        itemList[i] = result;
+      }
+    }
+
+    render();
+  } catch (error) {
     alert('unable to update text');
   }
-
-  const total = document.getElementById('total-sum');
-  const matchedTotal = total.innerText.match(itemSum);
-  let totalSum = Number(matchedTotal);
-  totalSum += Number(changedCost);
-  total.innerText = totalSum; 
-
-  render();
 }
 
 const cancelItemEditing = (id, lastWhere, lastWhen, lastSum) => {
-  const parent = document.getElementById(`container-${id}`);
-  if (parent === null) {
+  const myCostsList = document.getElementById(`list-with-costs`);
+  const container = document.getElementById(`container-${id}`);
+  const containerInfo = document.getElementById(`container-info-${id}`);
+  const buttons = document.getElementById(`buttons-${id}`);
+  const replaceWhere = document.getElementById(`replace-where-${id}`);
+  const replaceWhen = document.getElementById(`replace-when-${id}`);
+  const replaceCost = document.getElementById(`replace-cost-${id}`);
+  
+  if (myCostsList === null) {
     return;
   }
 
-  const writeWhere = document.createElement('p');
-  const writeWhen = document.createElement('p');
+  const where = document.createElement('p');
+  const when = document.createElement('p');
   const sum = document.createElement('p');
 
-  const inputs = parent.getElementsByTagName('input');
-  const buttons = parent.getElementsByTagName('button');
+  [where.innerText, when.innerText, sum.innerText] = [lastWhere, lastWhen, lastSum]; 
 
-  [writeWhere.innerText, writeWhen.innerText, sum.innerText] = [lastWhere, lastWhen, lastSum]; 
+  container.replaceChild(where, replaceWhere);
+  containerInfo.replaceChild(when, replaceWhen);
+  containerInfo.replaceChild(sum, replaceCost);
+  containerInfo.removeChild(buttons);
 
-  parent.removeChild(inputs[2]);
-  parent.removeChild(inputs[1]);
-  parent.removeChild(inputs[0]);
-  parent.removeChild(buttons[1]);
-  parent.removeChild(buttons[0]);
-  
+  where.className = 'write-where';
+  when.className = 'write-when';
+  sum.className = 'sum';
 
-  parent.appendChild(writeWhere);
-  parent.appendChild(writeWhen);
-  parent.appendChild(sum);
-
-  [writeWhere.className, writeWhen.className, sum.className] = ['write-where', 'write-when', 'sum'];
-
-  addButtons(id, lastWhere, lastWhen, lastSum);
+  where.id = `where-${id}`;
+  when.id = `when-${id}`;
+  sum.id = `sum-${id}`;
+ 
+  addButtons(id);
 }
 
-const addButtons = (id, lastWhere, lastWhen, lastSum) => {
-  const parent = document.getElementById(`container-${id}`);
+const addButtons = (id) => {
+  const containerInfo = document.getElementById(`container-info-${id}`);
   if (parent === null) {
     return;
   }
+
+  const buttons = document.createElement('div');
 
   const editButton = document.createElement('button');
   const deleteButton = document.createElement('button');
   const editImg = document.createElement('img');
   const deleteImg = document.createElement('img');
 
-  parent.appendChild(editButton);
-  parent.appendChild(deleteButton);
+  containerInfo.appendChild(buttons);
+  
+  buttons.appendChild(editButton);
+  buttons.appendChild(deleteButton);
   editButton.appendChild(editImg);
   deleteButton.appendChild(deleteImg);
+  
+  buttons.id = `buttons-${id}`;
+  buttons.className = 'list__buttons';
 
-  editButton.className = 'list__edit';
-  deleteButton.className = 'list__delete';
+  editButton.className = 'list__buttons-edit';
+  editButton.id = `edit-button-${id}`;
+
+  deleteButton.className = 'list__buttons-delete';
+  deleteButton.id = `delete-button-${id}`;
 
   editImg.src = 'img/edit.svg';
   deleteImg.src = 'img/delete.svg';
 
   editButton.onclick = () => {
-    editItem(id, lastWhere, lastWhen, lastSum);
+    editItem(id);
   }
 
   deleteButton.onclick = () => {
